Put the Explorer in a separate file.
